Refactor iMessage provider into focused modules
